GetImage method extraction for PR #3
