+++ WBooks/Models/AuthUser.swift
@@ -11,6 +11,6 @@ import Networking
 
 struct AuthUser: AuthenticableUser {
     // swiftlint:disable:next line_length
-    let sessionToken: String? = "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoxNDksInZlcmlmaWNhdGlvbl9jb2RlIjoiZkNmQWJ1bmpfS0xmVkp6cjhIeWQ3dHc1VlJSa1ZNQi1iNFhZWnlHd0pYc3NMenpUVm1uY3RHLTVGcFBxRHhocyIsInJlbmV3X2lkIjoiVHk0QnF6c0ZCdExLclp5eWRrMm1Lem02d0dFNzh3NzNBMXNGcS1qZjZ4ejF2RXNXVjd6ei14eFBGa0dGTkJIeCIsIm1heGltdW1fdXNlZnVsX2RhdGUiOjE1MjU2Mjk1NjYsImV4cGlyYXRpb25fZGF0ZSI6MTUyMzIxMDM2Niwid2FybmluZ19leHBpcmF0aW9uX2RhdGUiOjE1MjMwNTU1NjZ9.mIco1WYlnm4t58IdTTGjh6VlkK76WjlH3SY3L5En4WU"
+    let sessionToken: String? = "Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoxNDksInZlcmlmaWNhdGlvbl9jb2RlIjoiZkNmQWJ1bmpfS0xmVkp6cjhIeWQ3dHc1VlJSa1ZNQi1iNFhZWnlHd0pYc3NMenpUVm1uY3RHLTVGcFBxRHhocyIsInJlbmV3X2lkIjoidWs5aldwQVhzR3pvNVhlRHU5eml0XzRoNTE3Ukt4alVrX0JGZHlLUy1takU2eDJveFlzRGNOQ2d5SnNfZVpTciIsIm1heGltdW1fdXNlZnVsX2RhdGUiOjE1MjU4OTEwMDAsImV4cGlyYXRpb25fZGF0ZSI6MTUyMzQ3MTgwMCwid2FybmluZ19leHBpcmF0aW9uX2RhdGUiOjE1MjMzMTcwMDB9.c2OHAZKP0NQO51hVs6H8i12co8fbC-DXK89OHxfbuMc"
     
 }
