+++ WBooks/Models/CommentViewModel.swift
@@ -10,7 +10,7 @@ import Foundation
 import UIKit
 import Utils
 
-class CommentViewModel {
+class CommentViewModel: FetchedImageViewModelType {
     fileprivate var _comment: Comment
 
     init(comment: Comment) {
@@ -33,27 +33,7 @@ class CommentViewModel {
         return _comment.content
     }
 
-    private var _image: UIImage?
-
-    func getImage(closure: @escaping (UIImage) -> Void) {
-        guard let url = user.imageURL else { return }
-
-        if _image != .none {
-            closure(self._image!)
-        } else {
-            let imageFetcher = ImageFetcher()
-
-            imageFetcher.fetchImage(url).startWithResult { result in
-                switch result {
-                case .success(let image):
-                    self._image = image
-                    DispatchQueue.main.async {
-                        closure(image)
-                    }
-                case .failure(let error):
-                    print(error)
-                }
-            }
-        }
+    var imageURL: URL? {
+        return user.imageURL
     }
 }
